requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
selenium>=4.0.0
gTTS>=2.2.0
numpy>=1.19.0
requests>=2.25.0
python-dotenv>=0.19.0
webdriver_manager>=3.5.0 